Reorgnisation des routes

+++ routes/web.php
@@ -1,6 +1,5 @@
 <?php
 
-use App\Models\EvenementUser;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\RoleController;
@@ -14,22 +13,13 @@ Route::get('/', [EvenementController::class, 'accueil'])->name('accueil');
 Route::get('/evenement', [EvenementController::class, 'evenement'])->name('evenement');
 Route::get('/evenement-detail', [EvenementController::class, 'evenementDetail'])->name('evenement-detail');
 
-Route::middleware(['auth'])->group(function () {
-    Route::get('/reservations', [EvenementUserController::class, 'index'])->name('reservations.index');
-    Route::get('/creation', [EvenementController::class, 'create'])->name('creation');
-    Route::post('/creation', [EvenementController::class, 'creation'])->name('creation.store');
-});
-Route::get('/evenement/{id}', [EvenementController::class, 'evenementDetail'])->name('evenement.detail')->where('id', '[0-9]');
-Route::post('/evenement/{id}/reserver', [EvenementController::class, 'reserver'])->name('evenement.reserver')->where('id', '[0-9]');
-Route::get('/mes-evenements', [EvenementController::class, 'mesEvenements'])->name('mes.evenements');
-
-
-// Routes pour la création et gestion des événements')->where('id', '[0-9]')
-Route::resource('evenements', EvenementController::class);
+// Routes pour la création et gestion des événements
+Route::resource('evenements', EvenementController::class)->except(['create', 'store', 'edit', 'update']);
+Route::get('/creation', [EvenementController::class, 'create'])->name('creation');
+Route::post('/creation', [EvenementController::class, 'creation'])->name('creation.store');
 Route::put('/modifier/{id}', [EvenementController::class, 'modifier'])->name('modifier')->where('id', '[0-9]');
 Route::delete('/supprimer/{id}', [EvenementController::class, 'supprimer'])->name('supprimer')->where('id', '[0-9]');
 Route::get('/evenements/{id}', [EvenementController::class, 'show'])->name('evenements.show')->where('id', '[0-9]');
-
 Route::get('/mes-evenements', [EvenementController::class, 'mesEvenements'])->name('mes-evenements');
 
 // Routes pour les utilisateurs d'événements
@@ -48,49 +38,35 @@ Route::group(['prefix' => 'auth'], function () {
     Route::get('logout', [AuthController::class, 'logout'])->name('logout');
 });
 
-
-
+// Routes spécifiques pour les utilisateurs authentifiés
 Route::middleware(['auth'])->group(function () {
+    // Routes pour la gestion des rôles
+    Route::resource('roles', RoleController::class)->except(['destroy', 'edit', 'update']);
+    Route::delete('roles/{id}', [RoleController::class, 'destroy'])->name('roles.destroy')->middleware('permission:delete role');
+    Route::get('roles/{role}/edit', [RoleController::class, 'edit'])->name('roles.edit')->middleware('permission:update role');
+    Route::put('roles/{role}', [RoleController::class, 'update'])->name('roles.update')->middleware('permission:update role');
+    Route::get('roles/{id}/give-permissions', [RoleController::class, 'addPermissionToRole'])->name('role.permissions')->middleware('permission:add permission');
+    Route::put('roles/{id}/give-permissions', [RoleController::class, 'givePermissionToRole'])->name('role.permissions.update')->middleware('permission:add permission');
 
-    Route::controller(RoleController::class)->group(function () {
-        // Routes de ressource sans destroy, edit, update
-        Route::resource('roles', RoleController::class)->except(['destroy', 'edit', 'update']);
+    // Routes pour la gestion des permissions
+    Route::resource('permissions', PermissionController::class)->except(['destroy', 'edit', 'update']);
+    Route::delete('permissions/{id}', [PermissionController::class, 'destroy'])->name('permissions.destroy')->middleware('permission:delete permission');
+    Route::get('permissions/{permission}/edit', [PermissionController::class, 'edit'])->name('permissions.edit')->middleware('permission:update permission');
+    Route::put('permissions/{permission}', [PermissionController::class, 'update'])->name('permissions.update')->middleware('permission:update permission');
 
-        // Route pour la suppression des rôles avec middleware
-        Route::delete('roles/{id}', 'destroy')->name('roles.destroy')->middleware('permission:delete role');
+    // Routes pour l'administration des utilisateurs
+    Route::resource('users', UserController::class)->except(['create', 'store']);
+    Route::get('/dashboard', [UserController::class, 'dashboardAdmin'])->name('dashboard.admin');
+    Route::get('/profil-admin', [UserController::class, 'profilAdmin'])->name('profil.admin');
+    Route::post('/user/{id}/validate', [UserController::class, 'validateAccount'])->name('user.validate');
+    Route::post('/user/{id}/invalidate', [UserController::class, 'invalidateAccount'])->name('user.invalidate');
+    Route::post('/user/{id}/activate', [UserController::class, 'activateAccount'])->name('user.activate');
+    Route::post('/user/{id}/deactivate', [UserController::class, 'deactivateAccount'])->name('user.deactivate');
 
-        // Routes pour l'édition et la mise à jour des rôles avec middleware
-        Route::get('roles/{role}/edit', 'edit')->name('roles.edit')->middleware('permission:update role');
-        Route::put('roles/{role}', 'update')->name('roles.update')->middleware('permission:update role');
-
-        // Routes spécifiques pour les permissions associées aux rôles avec middleware
-        Route::get('roles/{id}/give-permissions', 'addPermissionToRole')->name('role.permissions')->middleware('permission:add permission');
-        Route::put('roles/{id}/give-permissions', 'givePermissionToRole')->name('role.permissions.update')->middleware('permission:add permission');
-    });
-
-
-    Route::controller(PermissionController::class)->group(function () {
-        // Routes de ressource sans destroy, edit, update
-        Route::resource('permissions', PermissionController::class)->except(['destroy', 'edit', 'update']);
-
-        // Route pour la suppression des permissions avec middleware
-        Route::delete('permissions/{id}', 'destroy')->name('permissions.destroy')->middleware('permission:delete permission');
-
-        // Routes pour l'édition et la mise à jour des permissions avec middleware
-        Route::get('permissions/{permission}/edit', 'edit')->name('permissions.edit')->middleware('permission:update permission');
-        Route::put('permissions/{permission}', 'update')->name('permissions.update')->middleware('permission:update permission');
-    });
-
-    Route::controller(UserController::class)->group(function () {
-        // Routes de ressource sans create et store
-        Route::resource('users', UserController::class)->except(['create', 'store']);
-
-        // Routes supplémentaires pour l'administration des utilisateurs
-        Route::get('/dashboard', 'dashboardAdmin')->name('dashboard.admin');
-        Route::get('/profil-admin', 'profilAdmin')->name('profil.admin');
-        Route::post('/user/{id}/validate', 'validateAccount')->name('user.validate');
-        Route::post('/user/{id}/invalidate', 'invalidateAccount')->name('user.invalidate');
-        Route::post('/user/{id}/activate', 'activateAccount')->name('user.activate');
-        Route::post('/user/{id}/deactivate', 'deactivateAccount')->name('user.deactivate');
-    });
+    // Route pour les réservations
+    Route::get('/reservations', [EvenementUserController::class, 'index'])->name('reservations.index');
 });
+
+// Routes supplémentaires
+Route::get('/evenement/{id}', [EvenementController::class, 'evenementDetail'])->name('evenement.detail')->where('id', '[0-9]');
+Route::post('/evenement/{id}/reserver', [EvenementController::class, 'reserver'])->name('evenement.reserver')->where('id', '[0-9]');
